refactor: change login mgr export path.
